Don't just add sub-items to the currently displayed list
This *might, sometimes* correctly add sub-items to the list where the parent item originated. More work is needed here.
The current code uses the first list that the parent item was on. Logically, this should work, but the ManyToManyField isn't sorted by "order added".
Since the field is sorted instead by primary key, if the currently displayed list has a lower PK than the originating list of the parent, it'll still get the list wrong.