Completely remove lists/quick-access-list-button.html
This moves functionality into lists/list-item.html, and re-renders the whole list every time a quick access button is pressed. Only the item being acted on is replaced in the DOM, to avoid wildly scrolling the page around. This re-rendering avoids a visual inconsistency where a list item would appear to have been removed from the current list, when in fact it would reappear when the page was reloaded.